Adding basic game logic (#2)
* added SuperTTT model

* added migrations

* Added lobby to Game, deleted default and new player type

* created migrations

* Added winner field and increase board size
Increased size of inner arrays to save the winner of the inner boards

* removed lobby attribute and added create game function

* added gameType attribute to lobby

* created game serializers and added game creations to views

* Added start game as a model method

* changed board items to integers

* Added views for the game

* testing game start

* added superttt players

* adding make move, its schema and permissions

* fixing object permissions checks

* IsInTurn permission fixed

* Register player moves

* test moves

* fix lobby attribute name changed

* simple game implemented

* deleted unnecesary atribute and migrations

* testing happy path and fix saving error

* Added superttt constraints

* update test response

+++ lobbies/views.py
@@ -7,6 +7,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
 import coreapi
 import coreschema
+from superttt.models import SuperTTT, SuperTTTPlayerSymbol
 
 lobby_population_schema = schemas.ManualSchema(
     fields=[
@@ -35,6 +36,7 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     user = self.request.user
     if lobby.players.contains(user):
       return Response(data={
@@ -58,6 +60,7 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     user = self.request.user
     if lobby.owner == user:
       lobby.delete()
@@ -77,14 +80,13 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     if lobby.players.count() != 2:
       return Response(data={
           'detail': 'Lobby must be full, 2 player needed for this game.'
         },
         status=status.HTTP_406_NOT_ACCEPTABLE
       )
-    lobby.started = True
-    lobby.save()
+    lobby.start_game()
     lobby_serializer = LobbySerializer(lobby)
-    # Falta lógica de inicio de juego
     return Response(data=lobby_serializer.data, status=status.HTTP_202_ACCEPTED)
